[ALPS03330660] DISP: Fix coverity issue
[Detail]
Fix coverity issue

MTK-Commit-Id: 90f99283cd89f97d1ff57592ef386e00babcc008

CR-Id: ALPS03330660
Feature: Others

+++ drivers/misc/mediatek/video/mt6765/videox/debug.h
@@ -21,7 +21,7 @@ void DBG_Deinit(void);
 extern int bypass_blank;
 extern int lcm_mode_status;
 extern int layer_layout_allow_non_continuous;
-extern unsigned int idle_check_interval;
+extern unsigned long long idle_check_interval;
 
 #include "disp_session.h"
 #include "ddp_info.h"
